fix perfil, falla comp aboutme
