+++ client-benkyo-devs/src/components/perfil/info-perfil/ProfileInfo.jsx
@@ -1,0 +1,11 @@
+const ProfileInfo = {
+    username: 'Carol',
+    about: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, voluptatum. Quisquam, voluptatum. Quisquam, voluptatum. Quisquam, voluptatum. Quisquam, voluptatum. Quisquam, voluptatum.',
+    avatar: 'https://images.unsplash.com/photo-1621574539436-4b0b0b0b0b0b?ixid=MnwxMjA3fDB8MHxzZWFyY2h8Mnx8YmVua3lvJTIwY2Fyb2wlMjBhdmF0YXJ8ZW58MHx8MHx8&ixlib=rb-1.2.1&w=1000&q=80',
+    points: {
+      easy: 101,
+      medium: 102,
+      hard: 103
+    }
+  }
+  export default ProfileInfo;
